chore: move v1 files
